Plotting Average Price of Every State

This function takes a `state' object to plot the average price histogram for the input state, and this value can be easily found from the graph by highlighting the selected state.

Usage

price_plot(state_name)

Arguments

state_name

character. Every state of US from HoneyProduction dataset.

Value

A bar chart showing the average price of honey in each state, with selected states highlighted and others grayed out, meanwhile each state's price data displayed on its own bar.

Examples

price_plot("AL")

price_plot(state_name = "AL")