Plotting Total Production Trend From 1998 to 2012

A line graph showing total honey production trend by states in the US over time, which is a big picture of the honey production data.

The output is the line graph of the honey production value for all states, regardless of whether argument has an input or not.

Usage

production_plot(allstates)

Arguments

allstates

character. Every state of US from HoneyProduction dataset.

Value

A line graph containing the curve of the change in the total honey production of each state from 1998 to 2012.

Examples

production_plot()

production_plot(allstates = )